fix(codeql): break runtime-bridge import cycle and document empty except
CodeQL flagged a module-level cyclic import between
backends/_runtime_bridge.py and backends/critic_agent.py: the bridge
referenced critic_agent.RuntimeCall while critic_agent imported
invoke_runtime_cli from the bridge, so RuntimeCall / invoke_runtime_cli
could be undefined depending on import order. robustness_agent also
imported RuntimeCall from critic_agent, leaking critic internals.

Move the shared RuntimeCall dataclass and RuntimeCaller alias into
_runtime_bridge.py (the bridge that owns the runtime.cli invocation),
widening phase to str and making review_path optional so both the critic
(prepare-review / commit-review) and robustness (tick) backends share one
definition. critic_agent and robustness_agent now import them from the
bridge, making the dependency one-directional and removing the cycle.

Also add an explanatory comment to the best-effort rmdir() except clause
in session_paths.py stale-workdir cleanup (tolerates non-empty dirs and
transient FS errors; the outer rmdir / next sweep retries).
